refac(ipipeline): rename the packages and modules
