--- ipipeline/control/__init__.py
@@ -1,7 +1,0 @@
-"""Package initialization procedures.
-
-The control package provides components that control the execution of 
-the user tasks.
-"""
-
-from ipipeline.control.execution import BaseExecutor, SequentialExecutor
