+++ ipipeline/execution/__init__.py
@@ -1,0 +1,6 @@
+"""Package initialization procedures.
+
+The execution package provides components to execute the tasks.
+"""
+
+from ipipeline.execution.executors import BaseExecutor, SequentialExecutor
