--- ipipeline/util/__init__.py
@@ -1,5 +1,0 @@
-"""Package initialization procedures.
-
-The util package provides components that are used in different modules 
-as utilities.
-"""
