+++ ipipeline/utils/__init__.py
@@ -1,0 +1,4 @@
+"""Package initialization procedures.
+
+The utils package provides components to use in different modules.
+"""
